+++ KeyboardKit/KeyInputStrings.swift
@@ -15,9 +15,10 @@ extension String {
     static let leftArrow = UIKeyCommand.inputLeftArrow
     static let rightArrow = UIKeyCommand.inputRightArrow
 
-    // TODO: Change these to the public constants once Xcode 11.4 is out of beta.
-    static let pageUp = "UIKeyInputPageUp"
-    static let pageDown = "UIKeyInputPageDown"
-    static let home = "UIKeyInputHome"
-    static let end = "UIKeyInputEnd"
+    static let pageUp = UIKeyCommand.inputPageUp
+    static let pageDown = UIKeyCommand.inputPageDown
+    // Home and End work as early as 13.3 and possibly 13.2. However the main reason for the fallback is that propagating
+    // the limited availability is annoying. There is no negative impact on old versions. The input just won’t be matched.
+    static let home: String = { if #available(iOS 13.4, *) { return UIKeyCommand.inputHome } else { return "UIKeyInputHome" } }()
+    static let end: String = { if #available(iOS 13.4, *) { return UIKeyCommand.inputEnd } else { return "UIKeyInputEnd" } }()
 }