+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # KeyboardKit change log
 
 - 2020-04-26
+    - Breaking change: Xcode 11.4 is now required.
     - Breaking change: KeyboardKit scroll views will no longer send their delegate `scrollViewDidEndScrollingAnimation(_:)` after keyboard-driven animations end. Instead, the scroll view’s delegate can conform to `KeyboardScrollingDelegate` to receive callbacks when keyboard-driven scrolling animations start or finish.
 - 2020-04-05
     - Fixes scroll views with `isPagingEnabled` set ending up off page boundaries if starting keyboard scrolling while the scrolling view was in-between pages while decelerating from touch scrolling.